+++ src/scanner_engine.cpp
@@ -246,9 +246,9 @@ namespace Scanner {
 
             auto now = std::chrono::steady_clock::now();
             auto elapsed_seconds = std::chrono::duration_cast<std::chrono::seconds>(now - current_stats.start_time).count();
-            int hours = elapsed_seconds / 3600;
-            int minutes = (elapsed_seconds % 3600) / 60;
-            int seconds = elapsed_seconds % 60;
+            int hours = static_cast<int>(elapsed_seconds / 3600);
+            int minutes = static_cast<int>((elapsed_seconds % 3600) / 60);
+            int seconds = static_cast<int>(elapsed_seconds % 60);
             std::string time_str = std::to_string(hours) + "h " + std::to_string(minutes) + "m " + std::to_string(seconds) + "s";
 
             std::cout << Config::current_time() << "Checkpoint saved at key: " << current_stats.current_position.to_hex()
